backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
python-multipart
sqlalchemy
psycopg2-binary
python-dotenv
pydantic[email]
google-auth
google-auth-oauthlib
httpx
